refactor(report): move assets to separate resources, decompose build_html, and decouple keying
