+++ tests/orchestration/test_execute_run.py
@@ -1338,7 +1338,8 @@ def test_run_report_phase_builds_report_from_current_clip_artifacts(
     report_data = captured["report_data"]
     assert artifacts.report_path == expected_report_path
     assert report_data.frames == [7, 11]
-    assert report_data.screenshots == artifacts.screenshots_by_label
+    assert report_data.clips[0].screenshots == artifacts.screenshots_by_label["Reference"]
+    assert report_data.clips[1].screenshots == artifacts.screenshots_by_label["Encode 1"]
     assert report_data.metadata == metadata
     assert report_data.slowpics_url == "https://slow.pics/c/collateral"
     assert [(clip.name, clip.frame_count) for clip in report_data.clips] == [
